Added CI so that the component can be built when it is changed.
The CI components have been updated so that headers amd libraries
can be built from the source. This is the first attempt to get this
working for the general components. In theory it should be reusable
for all the other components, but that might take a few days to get
right as there are a lot of components and many will surely have
different behaviour.

In local tests on MacOS, it seems to work. We'll see how it goes
in the CI runs.